'暂存'

+++ push/types.go
@@ -31,7 +31,7 @@ type (
     pushReq struct {
         FromUserId      string                `json:"From_Account"`    // （选填）消息推送方帐号
         Condition       *PushCondition        `json:"Condition"`       // （选填）推送条件
-        MsgRandom       int64                 `json:"MsgRandom"`       // （必填）消息随机数，由随机函数产生
+        MsgRandom       uint32                `json:"MsgRandom"`       // （必填）消息随机数，由随机函数产生
         MsgBody         []types.MsgBody       `json:"MsgBody"`         // （必填）消息内容
         MsgLifeTime     int                   `json:"MsgLifeTime"`     // （选填）消息离线存储时间，单位秒，最多保存7天（604800秒）。默认为0，表示不离线存储
         OfflinePushInfo types.OfflinePushInfo `json:"OfflinePushInfo"` // （选填）离线推送信息配置
